Modularize Views
Separated the ScheduledEntry table an DayNight tables into their own better structured directories and out of the main file.
